change must_eat

+++ philo/includes/s_philo.h
@@ -6,7 +6,7 @@
 /*   By: coleta <coleta@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/18 17:40:37 by coleta            #+#    #+#             */
-/*   Updated: 2022/06/08 20:57:15 by coleta           ###   ########.fr       */
+/*   Updated: 2022/06/08 21:39:21 by coleta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct s_philo
 	size_t			live;
 	size_t			id;
 	struct s_date	*date;
-	int				must_eat;
+	long			must_eat;
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	*right_fork;
 }t_philo;
